Convert site to Eleventy
This is a massive architecture shift that replaces the ageing Middleman
build system with Eleventy (11ty.io). To facilitate this, we also
convert all templates to Liquid, stylesheets to raw CSS, and scripts to
raw JS.

This is a modernization process that means the site can actually be
developed again, and isn't bogged down by an unnecessarily complex Ruby
build environment.