Re-add PT Sans import temporarily
I missed a bunch of the font-family properties because apparently it was
injected all over the place by the original Sass, and currently removing
PT Sans entirely breaks some of the layout things due to B612's
character width change.

So, for now just add it back to ensure things render correctly still,
until we can refactor the layout a bit more.