Move site back to Middleman
Eleventy was enticing for its lightness, but ultimately that "lightness"
turned out to just be a lack of useful features. Just trying to get CSS
to minify seemed to basically be an impossible problem without bringing
in a giant separate Gulp pipeline, which was just such a ballooning
scope that it was legitimately easier to just switch the entire site
back to Middleman.

This gives us a proven foundation, with Sass built right in, and all the
advanced features we could want, without a billion npm packages or
complex javascript build tools.