Refresh design around background image
The background image was looking a bit dated, and caused some contrast
issues with text over top sometimes, so it had to go.

Instead, we're now putting it above the content with the page title, as
it does provide a bit of contrast for an otherwise very white site, so
it's nice to keep it in some capacity.

This makes the content incredibly less straining to read, keeps the
general aesthetic, and makes the site look a lot more modern.