Update photos page to pull from data; add alt text
This updates the galleries to be pulled from YAML data instead of
hard-coded into the template. Additionally, this adds proper alt text
for screenreaders with descriptions of the photo, and captions for the
lightbox in some cases.

We also fix the weirdly-overkill use of `<article><figure>` in place of
a nice and simple unordered list.