Update lightbox library
This updates to "Fancybox" version 3 (previously we were on a very old
version). The new version works better on touch devices, looks much more
modern, and provides some very advanced features like gallery
navigation, slideshows, and zoom.

In tandem, since this is the only "real" JS on the site, we update
jQuery to version 3.