Convert site to Hugo
Middleman has proved difficult to maintain long-term as ruby versions
become unsupported and dependencies drift.

Hugo is just a static binary, so it is much less likely to have problems
as we can just run the older binary if we need to.

As a benefit, the site now builds _extremely_ fast, and is a bit more
modern as Hugo is "in-vogue" compared to Middleman which is a bit legacy
at this point.