Update PayPal button cmd
This broke at some point, maybe many years ago, but it just causes a
generic server error from PayPal now. Looking at the legacy docs for
this button type, it seems we probably want a "_donations" cmd, instead
of "_xclick" which is more checkout-targetted.

Changing this value does seem to load a proper donation page, so
hopefully that's all we need to change for now.